requirements.txt
discord.py[voice]
openai
python-dotenv
PyNaCl
discord-ext-voice-recv
